feat | sprint2 | FRB-70 | 환경 자동제어 기능 (센서값이 임계치±허용치 벗어날 때 로그 출력하던 기능에서 제어이력을 ControlLog에 저장하도록 기능 추가) | 윤다인

+++ src/main/java/com/factoreal/backend/entity/ControlLog.java
@@ -23,7 +23,7 @@ public class ControlLog {
     private String controlType;
 
     @Column(name = "control_val")
-    private Integer controlVal;
+    private Double controlVal;
 
     @Column(name = "control_stat")
     private Integer controlStat;
